refactor(outfitter): extract init output rendering helpers (OS-437)
